--- a/dhall.cabal
+++ b/dhall.cabal
@@ -182,9 +182,7 @@ Library
         directory                   >= 1.2.7.1  && < 1.4 ,
         exceptions                  >= 0.8.3    && < 0.11,
         filepath                    >= 1.4      && < 1.5 ,
-        hashable                                   < 1.3 ,
         haskeline                   >= 0.7.3.0  && < 0.8 ,
-        insert-ordered-containers   >= 0.2.1.0  && < 0.3 ,
         lens-family-core            >= 1.0.0    && < 1.3 ,
         megaparsec                  >= 7.0.0    && < 7.1 ,
         memory                      >= 0.14     && < 0.15,
@@ -221,10 +219,11 @@ Library
         Dhall.Hash,
         Dhall.Import,
         Dhall.Lint,
-        Dhall.Main
+        Dhall.Main,
+        Dhall.Map,
         Dhall.Parser,
         Dhall.Pretty,
-        Dhall.Repl
+        Dhall.Repl,
         Dhall.TH,
         Dhall.Tutorial,
         Dhall.TypeCheck
@@ -270,8 +269,6 @@ Test-Suite tasty
         containers                                     ,
         deepseq                   >= 1.2.0.1  && < 1.5 ,
         dhall                                          ,
-        hashable                                       ,
-        insert-ordered-containers == 0.2.1.0           ,
         prettyprinter                                  ,
         QuickCheck                >= 2.10     && < 2.13,
         quickcheck-instances      >= 0.3.12   && < 0.4 ,