--- a/src/Dhall.hs
+++ b/src/Dhall.hs
@@ -38,7 +38,7 @@ module Dhall
 
     -- * Types
     , Type(..)
-    , RecordType(..)
+    , RecordType
     , InputType(..)
     , Interpret(..)
     , InvalidType(..)
@@ -67,7 +67,7 @@ module Dhall
 
     , Inject(..)
     , inject
-    , RecordInputType(..)
+    , RecordInputType
     , inputFieldWith
     , inputField
     , inputRecord
@@ -92,6 +92,7 @@ import Data.Functor.Contravariant (Contravariant(..), (>$<))
 import Data.Functor.Contravariant.Divisible (Divisible(..), divided)
 import Data.Monoid ((<>))
 import Data.Scientific (Scientific)
+import Data.Semigroup
 import Data.Sequence (Seq)
 import Data.Text (Text)
 import Data.Typeable (Typeable)
@@ -711,7 +712,7 @@ unit = Type extractOut expectedOut
         | Data.Foldable.null fields = return ()
     extractOut _ = Nothing
 
-    expectedOut = Record mempty
+    expectedOut = Record Dhall.Map.empty
 
 {-| Decode a `String`
 
@@ -864,7 +865,7 @@ instance GenericInterpret V1 where
       where
         extract _ = Nothing
 
-        expected = Union mempty
+        expected = Union Dhall.Map.empty
 
 instance (Constructor c1, Constructor c2, GenericInterpret f1, GenericInterpret f2) => GenericInterpret (M1 C c1 f1 :+: M1 C c2 f2) where
     genericAutoWith options@(InterpretOptions {..}) = pure (Type {..})
@@ -1117,9 +1118,9 @@ instance Inject Double where
 instance Inject () where
     injectWith _ = InputType {..}
       where
-        embed = const (RecordLit mempty)
+        embed = const (RecordLit Dhall.Map.empty)
 
-        declared = Record mempty
+        declared = Record Dhall.Map.empty
 
 instance Inject a => Inject (Maybe a) where
     injectWith options = InputType embedOut declaredOut
@@ -1269,9 +1270,9 @@ instance (GenericInject f, GenericInject g) => GenericInject (f :*: g) where
 instance GenericInject U1 where
     genericInjectWith _ = pure (InputType {..})
       where
-        embed _ = RecordLit mempty
+        embed _ = RecordLit Dhall.Map.empty
 
-        declared = Record mempty
+        declared = Record Dhall.Map.empty
 
 instance (Selector s, Inject a) => GenericInject (M1 S s (K1 i a)) where
     genericInjectWith opts@(InterpretOptions {..}) = do
@@ -1327,7 +1328,7 @@ newtype RecordType a =
   RecordType
     ( Data.Functor.Product.Product
         ( Control.Applicative.Const
-            ( Dhall.Map.Map
+            ( MonoidMap
                 Text
                 ( Expr Src X )
             )
@@ -1340,10 +1341,19 @@ newtype RecordType a =
     )
   deriving (Functor, Applicative)
 
+newtype MonoidMap k v = MonoidMap (Dhall.Map.Map k v)
+
+instance Ord k => Data.Semigroup.Semigroup (MonoidMap k v) where
+    MonoidMap l <> MonoidMap r = MonoidMap (Dhall.Map.union l r)
+
+instance Ord k => Monoid (MonoidMap k v) where
+    mempty = MonoidMap Dhall.Map.empty
+
+    mappend = (<>)
 
 -- | Run a 'RecordType' parser to build a 'Type' parser.
 record :: RecordType a -> Dhall.Type a
-record ( RecordType ( Data.Functor.Product.Pair ( Control.Applicative.Const fields ) ( Data.Functor.Compose.Compose extractF ) ) ) =
+record ( RecordType ( Data.Functor.Product.Pair ( Control.Applicative.Const ( MonoidMap fields ) ) ( Data.Functor.Compose.Compose extractF ) ) ) =
   Type
     { extract =
         extractF
@@ -1367,9 +1377,11 @@ field key valueType =
     RecordType
       ( Data.Functor.Product.Pair
           ( Control.Applicative.Const
-              ( Dhall.Map.singleton
-                  key
-                  ( Dhall.expected valueType )
+              ( MonoidMap
+                  ( Dhall.Map.singleton
+                      key
+                      ( Dhall.expected valueType )
+                  )
               )
           )
           ( Data.Functor.Compose.Compose extractBody )
@@ -1440,10 +1452,12 @@ instance Contravariant RecordInputType where
 instance Divisible RecordInputType where
   divide f (RecordInputType bInputTypeRecord) (RecordInputType cInputTypeRecord) =
       RecordInputType
-    $ Dhall.Map.union
-      ((contramap $ fst . f) <$> bInputTypeRecord)
-      ((contramap $ snd . f) <$> cInputTypeRecord)
-  conquer = RecordInputType mempty
+          (Dhall.Map.union
+              ((contramap $ fst . f) <$> bInputTypeRecord)
+              ((contramap $ snd . f) <$> cInputTypeRecord)
+          )
+
+  conquer = RecordInputType Dhall.Map.empty
 
 inputFieldWith :: Text -> InputType a -> RecordInputType a
 inputFieldWith name inputType = RecordInputType $ Dhall.Map.singleton name inputType