--- a/src/Dhall.hs
+++ b/src/Dhall.hs
@@ -38,7 +38,7 @@ module Dhall
 
     -- * Types
     , Type(..)
-    , RecordType
+    , RecordType(..)
     , InputType(..)
     , Interpret(..)
     , InvalidType(..)
@@ -67,7 +67,7 @@ module Dhall
 
     , Inject(..)
     , inject
-    , RecordInputType
+    , RecordInputType(..)
     , inputFieldWith
     , inputField
     , inputRecord
@@ -91,7 +91,6 @@ import Control.Monad.Trans.State.Strict
 import Data.Functor.Contravariant (Contravariant(..), (>$<))
 import Data.Functor.Contravariant.Divisible (Divisible(..), divided)
 import Data.Scientific (Scientific)
-import Data.Semigroup
 import Data.Sequence (Seq)
 import Data.Text (Text)
 import Data.Typeable (Typeable)
@@ -711,7 +710,7 @@ unit = Type extractOut expectedOut
         | Data.Foldable.null fields = return ()
     extractOut _ = Nothing
 
-    expectedOut = Record Dhall.Map.empty
+    expectedOut = Record mempty
 
 {-| Decode a `String`
 
@@ -864,7 +863,7 @@ instance GenericInterpret V1 where
       where
         extract _ = Nothing
 
-        expected = Union Dhall.Map.empty
+        expected = Union mempty
 
 instance (Constructor c1, Constructor c2, GenericInterpret f1, GenericInterpret f2) => GenericInterpret (M1 C c1 f1 :+: M1 C c2 f2) where
     genericAutoWith options@(InterpretOptions {..}) = pure (Type {..})
@@ -1117,9 +1116,9 @@ instance Inject Double where
 instance Inject () where
     injectWith _ = InputType {..}
       where
-        embed = const (RecordLit Dhall.Map.empty)
+        embed = const (RecordLit mempty)
 
-        declared = Record Dhall.Map.empty
+        declared = Record mempty
 
 instance Inject a => Inject (Maybe a) where
     injectWith options = InputType embedOut declaredOut
@@ -1269,9 +1268,9 @@ instance (GenericInject f, GenericInject g) => GenericInject (f :*: g) where
 instance GenericInject U1 where
     genericInjectWith _ = pure (InputType {..})
       where
-        embed _ = RecordLit Dhall.Map.empty
+        embed _ = RecordLit mempty
 
-        declared = Record Dhall.Map.empty
+        declared = Record mempty
 
 instance (Selector s, Inject a) => GenericInject (M1 S s (K1 i a)) where
     genericInjectWith opts@(InterpretOptions {..}) = do
@@ -1327,7 +1326,7 @@ newtype RecordType a =
   RecordType
     ( Data.Functor.Product.Product
         ( Control.Applicative.Const
-            ( MonoidMap
+            ( Dhall.Map.Map
                 Text
                 ( Expr Src X )
             )
@@ -1340,19 +1339,10 @@ newtype RecordType a =
     )
   deriving (Functor, Applicative)
 
-newtype MonoidMap k v = MonoidMap (Dhall.Map.Map k v)
-
-instance Ord k => Data.Semigroup.Semigroup (MonoidMap k v) where
-    MonoidMap l <> MonoidMap r = MonoidMap (Dhall.Map.union l r)
-
-instance Ord k => Monoid (MonoidMap k v) where
-    mempty = MonoidMap Dhall.Map.empty
-
-    mappend = (<>)
 
 -- | Run a 'RecordType' parser to build a 'Type' parser.
 record :: RecordType a -> Dhall.Type a
-record ( RecordType ( Data.Functor.Product.Pair ( Control.Applicative.Const ( MonoidMap fields ) ) ( Data.Functor.Compose.Compose extractF ) ) ) =
+record ( RecordType ( Data.Functor.Product.Pair ( Control.Applicative.Const fields ) ( Data.Functor.Compose.Compose extractF ) ) ) =
   Type
     { extract =
         extractF
@@ -1376,11 +1366,9 @@ field key valueType =
     RecordType
       ( Data.Functor.Product.Pair
           ( Control.Applicative.Const
-              ( MonoidMap
-                  ( Dhall.Map.singleton
-                      key
-                      ( Dhall.expected valueType )
-                  )
+              ( Dhall.Map.singleton
+                  key
+                  ( Dhall.expected valueType )
               )
           )
           ( Data.Functor.Compose.Compose extractBody )
@@ -1451,12 +1439,10 @@ instance Contravariant RecordInputType where
 instance Divisible RecordInputType where
   divide f (RecordInputType bInputTypeRecord) (RecordInputType cInputTypeRecord) =
       RecordInputType
-          (Dhall.Map.union
-              ((contramap $ fst . f) <$> bInputTypeRecord)
-              ((contramap $ snd . f) <$> cInputTypeRecord)
-          )
-
-  conquer = RecordInputType Dhall.Map.empty
+    $ Dhall.Map.union
+      ((contramap $ fst . f) <$> bInputTypeRecord)
+      ((contramap $ snd . f) <$> cInputTypeRecord)
+  conquer = RecordInputType mempty
 
 inputFieldWith :: Text -> InputType a -> RecordInputType a
 inputFieldWith name inputType = RecordInputType $ Dhall.Map.singleton name inputType