… benchmark/deep-nested-large-record/Main.hs
… src/Dhall/Diff.hs
… src/Dhall/Freeze.hs
… src/Dhall/Import.hs
… src/Dhall/Map.hs
… src/Dhall/Parser/Combinators.hs
… src/Dhall/Parser/Expression.hs
… src/Dhall/Pretty/Internal.hs
… src/Dhall/TypeCheck.hs
… stack-lts-6.yaml
… tests/QuickCheck.hs
… tests/Regression.hs