This change drops the dependency on `insert-ordered-containers` in favor of a
new `Dhall.Map` module which implements a limited subset of the functionality we
need.  In addition to simplifying the dependency tree this also improves
the performance of the `deep-nested-large-record` benchmark by 2.4x:

Before:

```
time                 1.412 s    (1.388 s .. 1.436 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.398 s    (1.387 s .. 1.404 s)
std dev              10.02 ms   (2.502 ms .. 13.39 ms)
variance introduced by outliers: 19% (moderately inflated)
```

After:

```
time                 582.2 ms   (552.5 ms .. 628.4 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 565.9 ms   (560.3 ms .. 574.6 ms)
std dev              8.901 ms   (1.218 ms .. 11.75 ms)
variance introduced by outliers: 19% (moderately inflated)
```