@sellout 2018-10-01
@Gabriel439 Thanks! Yeah, @FintanH tried this and we saved a significant amount of time (50–60s) on one of our benchmarks. Our branch that just eliminates insert-order completely (plain Map, no sorting) shaves like another 50–60s, but at a loss of the normalized ordering (and `format` ordering). So, this branch gets about half the benefit with no feature compromise. There might be a bit more we can squeeze out of it, but this is a huge improvement.

@FintanH may have more details to share, since he’s been doing all our benchmarking.