@Gabriel439 2018-10-01
@sellout: Just out of curiosity: what is the total running time? (or, equivalently, what is the fractional performance improvement?)