@FintanH 2018-10-01
@Gabriel439 @sellout Hey! So I did a couple of runs today and noticed the speedup was even better. It went from ~190s to ~100s. I've been adding some tweaks like inlining and I'm seeing some promising feedback. Just witnessed a run that was 55s :grin: