@sellout 2018-10-01
So, actually, my numbers might be off – we have a _lot_ of benchmark numbers floating around at the moment.

But this branch gives us a 102s run. The previous time I see (including the `denote` change, but still using InsOrd) is 196s. However, that doesn’t have some of the other post 1.17 improvements on it, AFAIK, which probably explains the discrepancy from the 60s improvement I mentioned before.

Pinging @FintanH to verify.

Other than this Map stuff (which is rapidly improving) our other bottlenecks look a lot like
```
typeWithA.loop          Dhall.TypeCheck    src/Dhall/TypeCheck.hs:(98,5)-(794,49)   12.7    6.1
normalizeWith.loop.kts' Dhall.Core         src/Dhall/Core.hs:1682:9-28              10.8   12.7
normalizeWith.loop.kts' Dhall.Core         src/Dhall/Core.hs:1676:9-28               6.5    7.4
alphaNormalize.kts₁     Dhall.Core         src/Dhall/Core.hs:1240:5-35               5.6    6.8
...
alphaNormalize.kts₁     Dhall.Core         src/Dhall/Core.hs:1232:5-35               3.4    4.0
```
the normalization cases are for `Union` and `Record`, and I’m guessing that’s also where we’re spending time in `typeWithA`, but need to break it down first.