@FintanH 2018-10-01
To shed some more light on the context of the above profile. There are a lot of unions in our AST and these grow every time we use a constructor. We think that's where the bottle neck is but how to optimize that is the million dollar question.