@Gabriel439 2018-10-01
@FintanH: Give me a million dollars and I can fix it 😉 

In all seriousness, the next major step is just being more intelligent about where we are strict and lazy in normalization.  The logical conclusion of this is implementing the algorithm described in this book, since that algorithm prevents duplication of this sort of work:

https://dl.acm.org/citation.cfm?id=320040

Until that happens we might be able to play tricks by delaying expanding `constructors` fully until the end of normalization.