@jneira 2018-10-02
Hi, i've tried to build this pr with the `stack build --stack-yaml stack-lts-6.yaml` ([after updating it](https://github.com/dhall-lang/dhall-haskell/pull/613)), the lts needed to build etlas with dhall as dependency and i am getting several errors about importing `<>` operator in the Typecheck module:
```console
C:\dhall-haskell\src\Dhall\TypeCheck.hs:3622:12:
    Ambiguous occurrence ‘<>’ 
    It could refer to either ‘Data.Monoid.<>’, imported from ‘Data.Monoid’ at src\Dhall\TypeCheck.hs:28:21-24 
     or ‘Pretty.<>’,  imported from ‘Data.Semigroup’ at src\Dhall\TypeCheck.hs:30:24-36
```