@jneira 2018-10-02
For lts-6 the fix seems to be removing `import Data.Monoid ((<>))` in `Dhall.Typecheck` and `Dhall` modules. It seems ghc 7.10 is not able to see that `Data.Monoid` reexports the `Data.Semigroup` operator?