@jneira 2018-10-02
Another possible caveat: `dhall-to-cabal` is using another Map ops in the `CabalToDhall` module: [`Map.difference`](https://github.com/dhall-lang/dhall-to-cabal/blob/master/lib/CabalToDhall.hs#L171),[`Map.filter`](https://github.com/dhall-lang/dhall-to-cabal/blob/master/lib/CabalToDhall.hs#L167),[`Map.mapMaybe`](https://github.com/dhall-lang/dhall-to-cabal/blob/master/lib/CabalToDhall.hs#L444)