@jneira 2018-10-03
Performance of compile dhall-to-cabal files has improved significantly :fireworks: :
```console
C:\dhall-to-cabal>timer "dhall-master < dhall-to-cabal.dhall > null"
00:00:02,44
00:00:23,93

C:\dhall-to-cabal>timer "dhall-newmap < dhall-to-cabal.dhall > null"
00:00:36,72
00:00:48,08
```
The new profile shows that cost has moved to `Dhall.Parser.Combinators` (although the new `Dhall.Map` is the second one):
```
	   dhall +RTS -p -RTS

	total time  =       14.47 secs   (14467 ticks @ 1000 us, 1 processor)
	total alloc = 22,183,225,456 bytes  (excludes profiling overheads)

COST CENTRE            MODULE                   SRC                                       %time %alloc

>>=                    Dhall.Parser.Combinators src\Dhall\Parser\Combinators.hs:94:5-48    18.3    1.7
traverseWithKey_       Dhall.Map                src\Dhall\Map.hs:302:1-77                  11.7   22.2
fmap                   Dhall.Map                src\Dhall\Map.hs:73:3-39                    9.8   15.9
typeWithA.loop         Dhall.TypeCheck          src\Dhall\TypeCheck.hs:(98,5)-(788,49)      8.9    0.7
normalizeWith.loop     Dhall.Core               src\Dhall\Core.hs:(1482,2)-(1833,22)        6.3    8.5
normalizeWith          Dhall.Core               src\Dhall\Core.hs:(1480,1)-(1833,22)        5.3   18.2
normalize              Dhall.Core               src\Dhall\Core.hs:1372:1-41                 3.6    4.3
denote                 Dhall.Core               src\Dhall\Core.hs:(1398,1)-(1462,39)        3.2    3.1
typeWithA.loop.process Dhall.TypeCheck          src\Dhall\TypeCheck.hs:(534,13)-(539,85)    2.9    0.0
exprFromUncachedImport Dhall.Import             src\Dhall\Import.hs:(580,1)-(673,45)        2.8    0.3
<|>                    Text.Megaparsec.Internal Text\Megaparsec\Internal.hs:195:3-16        2.3    2.0
fmap                   Dhall.Context            src\Dhall\Context.hs:32:15-21               2.0    6.6
pure                   Dhall.Parser.Combinators src\Dhall\Parser\Combinators.hs:75:5-24     1.9    0.9
shift                  Dhall.Core               src\Dhall\Core.hs:(758,1)-(920,29)          1.8    1.9
fmap                   Dhall.Parser.Combinators src\Dhall\Parser\Combinators.hs:68:5-41     1.3    1.2
lookup                 Dhall.Context            src\Dhall\Context.hs:(61,1)-(68,33)         1.2    0.1
getParserState         Dhall.Parser.Combinators src\Dhall\Parser\Combinators.hs:155:5-58    1.1    0.7
```