@Gabriel439 2018-10-03
@jneira: Do those timings mean that this branch got slower compared to `master`?