@jneira 2018-10-03
@Gabriel439 no, no, sorry for the poor stats but master took ~21 seconds and new_map ~11 seconds