@jneira 2018-10-04
@Gabriel439 is there any plan for implement the Map operations used by dhall-to-cabal (difference, filter,...)