@Gabriel439 2018-10-06
@jneira: Alright, I've pushed a branch of `dhall-to-cabal` that builds against this branch of `dhall`: https://github.com/dhall-lang/dhall-to-cabal/compare/gabriel/latest_dhall

I'm just going to add some doctests for `Dhall.Map` before merging