@jneira 2018-10-09
@Gabriel439 Not sure if i should comment here but i've detected that performance of actual master version is slightly worse that the pr (at least [my branch created from it](https://github.com/eta-lang/dhall-haskell/tree/new_map)

```console
jneira@laptop MINGW64 /d/ws/eta/dhall/dhall-to-cabal
$ time dhall-newmap < dhall-to-cabal.dhall > nul

real    0m11.479s
user    0m0.000s
sys     0m0.031s

jneira@laptop MINGW64 /d/ws/eta/dhall/dhall-to-cabal
$ time dhall-master < dhall-to-cabal.dhall > nul

real    0m16.116s
user    0m0.000s
sys     0m0.015s
```
Those results match the benchmarks? 
cost of the branch is around 11 secs and master 16 :thinking: